@@ -5,7 +5,7 @@ version: 1.0.0+1
 publish_to: 'none'
 
 environment:
-  sdk: '>=2.11.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
   flutter: '>=2.0.0'
 
 dependencies:
