@@ -6,14 +6,14 @@ import '../loading_state_view_model.dart';
 import 'loading.dart';
 
 class ContainerWithLoading extends StatelessWidget {
-  ContainerWithLoading({Widget child}) : _child = child;
+  ContainerWithLoading({Widget? child}) : _child = child;
 
-  final Widget _child;
+  final Widget? _child;
 
   @override
   Widget build(BuildContext context) {
     return Stack(children: [
-      _child,
+      _child!,
       HookBuilder(builder: (context) {
         final state = useProvider(loadingStateProvider);
         return state.isLoading ? const Loading() : const SizedBox();
