@@ -3,20 +3,19 @@ import 'package:freezed_annotation/freezed_annotation.dart';
 import 'source.dart';
 
 part 'article.freezed.dart';
-
 part 'article.g.dart';
 
 @freezed
 abstract class Article with _$Article {
   factory Article({
-    Source source,
-    String author,
-    String title,
-    String description,
-    String url,
-    String urlToImage,
-    DateTime publishedAt,
-    String content,
+    Source? source,
+    String? author,
+    String? title,
+    String? description,
+    String? url,
+    String? urlToImage,
+    DateTime? publishedAt,
+    String? content,
   }) = _Article;
 
   factory Article.fromJson(Map<String, dynamic> json) =>
