@@ -10,10 +10,10 @@ _$_News _$_$_NewsFromJson(Map<String, dynamic> json) {
   return _$_News(
     status: json['status'] as String,
     totalResults: json['totalResults'] as int,
-    articles: (json['articles'] as List)
-        ?.map((e) =>
-            e == null ? null : Article.fromJson(e as Map<String, dynamic>))
-        ?.toList(),
+    articles: (json['articles'] as List<dynamic>?)
+            ?.map((e) => Article.fromJson(e as Map<String, dynamic>))
+            .toList() ??
+        [],
   );
 }
 
