@@ -1,7 +1,7 @@
 import 'package:firebase_auth/firebase_auth.dart';
 
 abstract class AuthDataSource {
-  Future<User> signIn();
+  Future<User?> signIn();
 
   Future<void> signOut();
 }
