@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
 abstract class ThemeRepository {
-  Future<ThemeMode> loadThemeMode();
+  Future<ThemeMode?> loadThemeMode();
 
-  Future<void> saveThemeMode(ThemeMode theme);
+  Future<void> saveThemeMode(ThemeMode? theme);
 }
