@@ -2,8 +2,8 @@ import 'package:flutter/widgets.dart';
 
 import '../../gen/assets.gen.dart';
 
-ImageProvider loadProfileImage(String imageUrl) {
+ImageProvider loadProfileImage(String? imageUrl) {
   return imageUrl == null || imageUrl.isEmpty
       ? Assets.images.iconPlaceholder
-      : Image.network(imageUrl);
+      : Image.network(imageUrl) as ImageProvider<Object>;
 }
