@@ -2,7 +2,7 @@ import 'package:flutter/widgets.dart';
 
 import '../../gen/assets.gen.dart';
 
-Widget networkImage(String url, {BoxFit fit}) {
+Widget networkImage(String? url, {BoxFit? fit}) {
   final placeholder = Assets.images.articlePlaceholder.image(fit: fit);
 
   if (url == null) {
